Make episodes have a random unique id to avoid collisions between runs

+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Game/CarlaEpisode.h
@@ -275,7 +275,7 @@ private:
     ElapsedGameTime += DeltaSeconds;
   }
 
-  const uint32 Id = 0u;
+  const uint64 Id = 0u;
 
   double ElapsedGameTime = 0.0;
